requirements.txt
asttokens==3.0.0
colorama==0.4.6
executing==2.2.0
icecream==2.1.4
pyaes==1.6.1
pyasn1==0.6.1
Pygments==2.19.1
python-dotenv==1.0.1
rsa==4.9
Telethon==1.39.0
